fix: create initial state in Tabs wrapper
