+++ src/organisms/tabs/list/TabList.tsx
@@ -2,7 +2,7 @@ import { Button } from '@atoms';
 import { ButtonView, IconName, OrientationMode } from '@enums';
 import { ChopLogicTabItem } from '@models';
 import { getClassName, moveFocusOnElementById } from '@utils';
-import { FC, KeyboardEvent, useEffect, useRef } from 'react';
+import { FC, KeyboardEvent, useEffect } from 'react';
 
 import { TabButton } from '../button/TabButton';
 import styles from './TabList.module.scss';
@@ -20,6 +20,7 @@ type ChopLogicTabListProps = {
   extendable?: boolean;
   onTabAdd?: () => void;
   extendedTabLabel?: string;
+  initialTabsCount: number;
 };
 
 export const TabList: FC<ChopLogicTabListProps> = ({
@@ -35,9 +36,9 @@ export const TabList: FC<ChopLogicTabListProps> = ({
   extendable = false,
   onTabAdd,
   extendedTabLabel = 'New Tab',
+  initialTabsCount,
 }) => {
   const listClass = getClassName([styles.tabList, { [styles.tabList__vertical]: mode === OrientationMode.Vertical }]);
-  const initialTabsCount = useRef(tabs.length);
 
   const handleListKeyDown = (e: KeyboardEvent<HTMLDivElement>) => {
     const currentFocusedTabIndex = tabIds.findIndex((id) => id === selectedTabId);
@@ -82,7 +83,7 @@ export const TabList: FC<ChopLogicTabListProps> = ({
 
   useEffect(() => {
     // Auto-select newly added tab
-    if (tabs.length > initialTabsCount.current) {
+    if (tabs.length > initialTabsCount) {
       onTabSelect(tabs[tabs.length - 1].id);
     }
   }, [tabs.length]);
